+++ src/components/Major/MajorSelect.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { MajorSearch, Modal } from 'components';
 import { subStr } from 'utils/subString';
-import { Arrows, OptionBox, SelectedOption, SelectedOption_M } from 'styles/ommon';
+import { Arrows, OptionBox, SelectedOption, SelectedOption_M } from 'styles/common';
 
 const MajorSelect = () => {
   const [searchParams] = useSearchParams();
